requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.7.1
passlib==1.7.4
openpyxl==3.1.5
